# api_go/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for SmartOrganizr.",
        "title": "SmartOrganizr API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/public/invitations/{token}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get a public club invitation by token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubInvitationPublicDto"
                        }
                    }
                }
            }
        },
        "/public/invitations/{token}/complete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Set password and complete a public invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Completion payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubInvitationCompleteDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v1/clubs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Create a new club (with the current user as admin)",
                "parameters": [
                    {
                        "description": "Club payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubDto"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/me/permissions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Get the current user's permissions inside a club",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubPermissionsDto"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/members": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "List members of a club",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClubMemberDto"
                            }
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/members/export": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Download club members as CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/members/import": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Invite club members from a CSV upload (multipart \"file\")",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubInviteResultDto"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/members/invite": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Invite users by email to a club",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Email payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubInvitePostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubInviteResultDto"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/members/{memberUserId}/role": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "Change a club member's role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member user ID",
                        "name": "memberUserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubMemberRolePatchDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v1/clubs/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clubs"
                ],
                "summary": "List clubs the current user belongs to",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClubDto"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/invitations/{token}/accept": {
            "post": {
                "tags": [
                    "invitations"
                ],
                "summary": "Authenticated accept of a club invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ClubDto": {
            "type": "object",
            "required": [
                "club_type",
                "confirmed_representative",
                "country",
                "dates_visible_for_all_members",
                "feedback_visibility",
                "house_number",
                "id",
                "location",
                "members_can_send_messages",
                "name",
                "postal_code",
                "reason_visibility",
                "street"
            ],
            "properties": {
                "club_type": {
                    "type": "string"
                },
                "confirmed_representative": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string"
                },
                "dates_visible_for_all_members": {
                    "type": "boolean"
                },
                "feedback_visibility": {
                    "type": "string"
                },
                "house_number": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "members_can_send_messages": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "reason_visibility": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "dto.ClubInvitationCompleteDto": {
            "type": "object",
            "required": [
                "confirm_password",
                "firstname",
                "lastname",
                "password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.ClubInvitationPublicDto": {
            "type": "object",
            "required": [
                "club_id",
                "club_name",
                "expires_at",
                "invited_email",
                "is_accepted",
                "is_expired",
                "token"
            ],
            "properties": {
                "club_id": {
                    "type": "string"
                },
                "club_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "invited_email": {
                    "type": "string"
                },
                "is_accepted": {
                    "type": "boolean"
                },
                "is_expired": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.ClubInvitePostDto": {
            "type": "object",
            "required": [
                "emails"
            ],
            "properties": {
                "emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.ClubMemberDto": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "role",
                "user_id",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.ClubPermissionsDto": {
            "type": "object",
            "required": [
                "can_invite_members",
                "can_manage_roles",
                "role",
                "section_write"
            ],
            "properties": {
                "can_invite_members": {
                    "type": "boolean"
                },
                "can_manage_roles": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "section_write": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                }
            }
        },
        "dto.ClubPostDto": {
            "type": "object",
            "required": [
                "club_type",
                "confirmed_representative",
                "country",
                "dates_visible_for_all_members",
                "feedback_visibility",
                "house_number",
                "location",
                "members_can_send_messages",
                "name",
                "postal_code",
                "reason_visibility",
                "street"
            ],
            "properties": {
                "club_type": {
                    "type": "string"
                },
                "confirmed_representative": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string"
                },
                "dates_visible_for_all_members": {
                    "type": "boolean"
                },
                "feedback_visibility": {
                    "type": "string"
                },
                "house_number": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "members_can_send_messages": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "reason_visibility": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
...
    }
}